requirements.txt
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
entrypoints==0.3
flake8==3.7.7
mccabe==0.6.1
Pillow==6.0.0
pkg-resources==0.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.1
sqlparse==0.3.0
